improved speed for exporting NIFTIs - Got the correct pixel coordinates for of each Channel on each slice
Instead of saving the figure, the axis is saved.
Since it is much faster, we might be able to avoid the need for parallel processing.
The pixel coordinates of the channels are acquired, however this needs to be improved.

For now this works with a specific size of a figure, a change of the axis ratios to [1,1,1] and a hardcoded x_coordinates scale (1.16)